+++ CHANGELOG.md
@@ -4,6 +4,8 @@ ADDED:
 
 CHANGED:
 
+- [#5](https://github.com/woohoolabs/harmony/issues/5): Return a Response object for all the middlewares
+
 FIXED:
 
 ## 2.2.0 - 2015-12-16